requirements.txt

fastapi==0.128.0
opencv-python==4.12.0.88
pillow==12.1.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
requests-toolbelt==1.0.0
torch==2.9.1
torchvision==0.24.1
ultralytics==8.2.30
ultralytics-thop==2.0.18
uvicorn==0.40.0

